+ProdutoService +ProdutoController

+++ src/main/java/rest/api/domain/model/entities/Produto.kt
@@ -6,4 +6,4 @@ import javax.persistence.GenerationType
 import javax.persistence.Id
 
 @Entity
-data class Produto(val nome: String, val tipo: String, @Id @GeneratedValue(strategy = GenerationType.AUTO) var id: Int? = null)
+data class Produto(val nome: String, val tipo: String, @Id @GeneratedValue(strategy = GenerationType.AUTO) var id: Long? = null)
